+++ packages/semantic-ui/src/PasswordWidget/PasswordWidget.js
@@ -27,10 +27,6 @@ function PasswordWidget(props) {
     uiSchema,
     formContext,
     options,
-    defaultSchemaProps: {
-      inverted: false,
-      fluid: true,
-    }
  });
   // eslint-disable-next-line no-shadow
   const _onChange = ({ target: { value } }) =>