+++ packages/semantic-ui/src/URLWidget/URLWidget.js
@@ -23,12 +23,9 @@ function URLWidget(props) {
     formContext,
   } = props;
   const semanticProps = getSemanticProps(
-    { formContext, options,
+    { formContext,
+      options,
       uiSchema,
-      defaultSchemaProps: {
-        fluid: true,
-        inverted: false,
-    }
   });
   // eslint-disable-next-line no-shadow
   const _onChange = ({ target: { value } }) =>