+++ CHANGELOG.md
@@ -20,9 +20,9 @@ should change the heading of the (upcoming) version to include a major version b
 # v3.3.0
 
 ## @rjsf/semantic-ui
-- "semantic-ui-react" updated to v1.2.1 (https://github.com/rjsf-team/react-jsonschema-form/pull/2590)
+- "semantic-ui-react" updated to v1.3.1 (https://github.com/rjsf-team/react-jsonschema-form/pull/2590)
 - fixed passing [semantic theme specific props](https://react-jsonschema-form.readthedocs.io/en/latest/api-reference/themes/semantic-ui/uiSchema/), by overiding defaults, with your own semantic props (https://github.com/rjsf-team/react-jsonschema-form/pull/2590)
-  
+
 # v3.2.0
 
 ## @rjsf/core